directory 정리, jpa + mysql 설정
