--- src/main/java/bis/stock/back/RestController.java
@@ -1,16 +1,0 @@
-package bis.stock.back;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-@Controller
-public class RestController {
-	@RequestMapping("/")
-	@ResponseBody
-	public String home(){
-
-		return "된다";
-
-	}
-}
