+++ src/main/java/bis/stock/back/domain/user/User.java
@@ -1,0 +1,20 @@
+package bis.stock.back.domain.user;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+// DB 테이블 생성 테스트용 엔티티 작성
+@NoArgsConstructor
+@Getter
+@Entity
+public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+}
