+++ src/main/resources/application.yml
@@ -1,0 +1,23 @@
+spring:
+  datasource: #database source 작성
+    url: > # localhost 부분 클라우드 ip로 변경. 하지만 배포할 때에는 localhost로 되어 있어야 접근할 수 있음.
+      jdbc:mysql://localhost:3306/stocking?
+      useSSL=false&characterEncoding=UTF-8&serverTimezone=UTC&
+      allowPublicKeyRetrieval=true&createDatabaseIfNotExist=true
+    # DB 사용자 아이디와 password 작성
+    username: root
+    password: bi1234!@
+    driver-class-name: com.mysql.cj.jdbc.Driver
+  jackson:
+    property-naming-strategy: SNAKE_CASE
+
+  jpa:
+    generate-ddl: true
+    database: mysql
+    hibernate:
+      ddl-auto: create-drop # application 실행 시점에 기존 ddl 정보를 다 지우고 다시 실행시킨다. 그대로 유지하려면 none
+    properties:
+      hibernate:
+        #        show_sql: true # jpa, hibernate가 생성하는 모든 sql이 sout으로 찍힌다
+        # 운영환경에서는 sout이 아닌 log로 나타내야 한다
+        format_sql: true
